UHF-3505: Test fixes, moved service channel migrate to own test class
