+++ src/Fixture/Unit.php
@@ -79,6 +79,7 @@ final class Unit extends FixtureBase {
         'accessibility_viewpoints' => '00:unknown,11:red,12:red,13:red,21:green,22:green,23:green,31:red,32:red,33:red,41:green,51:red,52:red,61:red',
         'created_time' => '2019-09-18T10:32:56',
         'modified_time' => '2021-11-23T00:32:37',
+        'services' => [],
         'connections' => [
           [
             'section_type' => 'OPENING_HOURS',
@@ -288,6 +289,7 @@ Vantaan sairaala on opetussairaala, joten potilaiden hoitoon osallistuu sek√§ l√
         'accessibility_viewpoints' => '00:unknown,11:green,12:green,13:green,21:green,22:green,23:green,31:green,32:green,33:green,41:green,51:red,52:red,61:red',
         'created_time' => '2021-10-11T11:03:55',
         'modified_time' => '2021-10-28T14:40:08',
+        'services' => [],
         'connections' => [],
         'ontologyword_details' => [
           [
@@ -447,6 +449,7 @@ Vantaan sairaala on opetussairaala, joten potilaiden hoitoon osallistuu sek√§ l√
         'accessibility_viewpoints' => '00:unknown,11:unknown,12:unknown,13:unknown,21:unknown,22:unknown,23:unknown,31:unknown,32:unknown,33:unknown,41:unknown,51:unknown,52:unknown,61:unknown',
         'created_time' => '2020-10-09T21:38:01',
         'modified_time' => '2021-11-10T00:32:39',
+        'services' => [],
         'connections' => [
           [
             'section_type' => 'OPENING_HOURS',
@@ -541,6 +544,12 @@ Vantaan sairaala on opetussairaala, joten potilaiden hoitoon osallistuu sek√§ l√
             'sentence_fi' => 'Sentence fi 2',
             'sentence_sv' => 'Sentence sv 2',
             'sentence_en' => 'Sentence en 2',
+          ],
+        ],
+        'services' => [
+          [
+            'unit_id' => 1,
+            'services' => [1 => 1],
           ],
         ],
         'connections' => [
@@ -650,6 +659,7 @@ Vantaan sairaala on opetussairaala, joten potilaiden hoitoon osallistuu sek√§ l√
         'accessibility_viewpoints' => '00:unknown,11:unknown,12:unknown,13:unknown,21:unknown,22:unknown,23:unknown,31:unknown,32:unknown,33:unknown,41:unknown,51:unknown,52:unknown,61:unknown',
         'created_time' => '2015-02-11T13:36:50',
         'modified_time' => '2021-11-10T00:32:35',
+        'services' => [],
         'connections' => [
           [
             'section_type' => 'OPENING_HOURS',
@@ -681,6 +691,7 @@ Vantaan sairaala on opetussairaala, joten potilaiden hoitoon osallistuu sek√§ l√
     foreach ($units as $unit) {
       $responses[] = new Response(200, [], json_encode($unit['accessibility_sentences']));
       $responses[] = new Response(200, [], json_encode($unit['connections']));
+      $responses[] = new Response(200, [], json_encode($unit['services']));
     }
     return $responses;
   }
